# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CoinSphere API
  version: 1.0.0
  description: >-
    The versioned HTTP and WebSocket contract for the CoinSphere modular
    monolith. This document is the source for the generated Go and TypeScript
    contract types.
servers:
  - url: /
    description: Same-origin deployment
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/reauth:
    post:
      tags: [auth]
      operationId: reauthenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReauthRequest'
      responses:
        '200':
          description: One-time token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReauthResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/me:
    get:
      tags: [me]
      operationId: getCurrentUser
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/home/meta:
    get:
      tags: [me]
      operationId: getHomeMeta
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/home/overview:
    get:
      tags: [me]
      operationId: getHomeOverview
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/data/news:
    get:
      tags: [data]
      operationId: listNews
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Keyword'
      responses:
        '200': { $ref: '#/components/responses/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [data]
      operationId: createNews
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/data/news/{newsId}:
    parameters:
      - $ref: '#/components/parameters/NewsId'
    put:
      tags: [data]
      operationId: updateNews
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [data]
      operationId: deleteNews
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/admin/users:
    get:
      tags: [admin]
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: id
          in: query
          schema: { type: integer, format: int64, minimum: 1 }
        - name: username
          in: query
          schema: { type: string }
        - name: gender
          in: query
          schema: { type: string }
        - name: phone
          in: query
          schema: { type: string }
        - name: email
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin]
      operationId: createUser
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [admin]
      operationId: updateUser
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [admin]
      operationId: deleteUser
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/uploads/avatars:
    post:
      tags: [admin]
      operationId: uploadAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar: { type: string, format: binary }
              required: [avatar]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/system/roles:
    get:
      tags: [admin]
      operationId: listRoles
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: id
          in: query
          schema: { type: integer, format: int64, minimum: 1 }
        - name: displayName
          in: query
          schema: { type: string }
        - name: code
          in: query
          schema: { type: string }
        - name: description
          in: query
          schema: { type: string }
        - name: isEnabled
          in: query
          schema: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin]
      operationId: createRole
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/roles/{roleId}:
    parameters:
      - $ref: '#/components/parameters/RoleId'
    put:
      tags: [admin]
      operationId: updateRole
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [admin]
      operationId: deleteRole
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/roles/{roleId}/permissions:
    parameters:
      - $ref: '#/components/parameters/RoleId'
    put:
      tags: [admin]
      operationId: saveRolePermissions
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/menus:
    get:
      tags: [admin]
      operationId: listMenus
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [admin]
      operationId: createMenu
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/menus/manage-tree:
    get:
      tags: [admin]
      operationId: getManageMenuTree
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/i18n-dictionaries:
    get:
      tags: [admin]
      operationId: getI18nDictionaries
      parameters:
        - name: scope
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/system/menus/{menuId}:
    parameters:
      - $ref: '#/components/parameters/MenuId'
    put:
      tags: [admin]
      operationId: updateMenu
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [admin]
      operationId: deleteMenu
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/menu-buttons:
    post:
      tags: [admin]
      operationId: createMenuButton
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/system/menu-buttons/{buttonId}:
    parameters:
      - $ref: '#/components/parameters/ButtonId'
    put:
      tags: [admin]
      operationId: updateMenuButton
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [admin]
      operationId: deleteMenuButton
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/overview:
    get:
      tags: [workflows]
      operationId: getWorkflowOverview
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/task-definitions:
    get:
      tags: [workflows]
      operationId: listTaskDefinitions
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/task-definitions/page:
    get:
      tags: [workflows]
      operationId: listTaskDefinitionPage
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Keyword'
      responses:
        '200': { $ref: '#/components/responses/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/task-definitions/{taskCode}/default-params:
    parameters:
      - $ref: '#/components/parameters/TaskCode'
    put:
      tags: [workflows]
      operationId: updateTaskDefinitionDefaults
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/node-definitions:
    get:
      tags: [workflows]
      operationId: listWorkflowNodeDefinitions
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/agent-options:
    get:
      tags: [workflows]
      operationId: listWorkflowAgentOptions
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows:
    get:
      tags: [workflows]
      operationId: listWorkflows
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [workflows]
      operationId: createWorkflow
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/validate:
    post:
      tags: [workflows]
      operationId: validateWorkflow
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
    get:
      tags: [workflows]
      operationId: getWorkflow
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      tags: [workflows]
      operationId: updateWorkflow
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: [workflows]
      operationId: deleteWorkflow
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}/activate:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
    post:
      tags: [workflows]
      operationId: activateWorkflow
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}/deactivate:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
    post:
      tags: [workflows]
      operationId: deactivateWorkflow
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}/runtime:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
    get:
      tags: [workflows]
      operationId: getWorkflowRuntime
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}/runtime/entries/{entryKey}:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
      - $ref: '#/components/parameters/EntryKey'
    patch:
      tags: [workflows]
      operationId: updateWorkflowRuntimeEntry
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}/runtime/entries/{entryKey}/rotate-secret:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
      - $ref: '#/components/parameters/EntryKey'
    post:
      tags: [workflows]
      operationId: rotateWorkflowSecret
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/workflows/{definitionId}/executions:
    parameters:
      - $ref: '#/components/parameters/DefinitionId'
    get:
      tags: [workflows]
      operationId: listWorkflowExecutions
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Keyword'
        - name: triggerType
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [workflows]
      operationId: runWorkflow
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody: { $ref: '#/components/requestBodies/JsonObject' }
      responses:
        '202': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/workflows/executions:
    get:
      tags: [workflows]
      operationId: listAllWorkflowExecutions
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: workflowDefinitionCode
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Keyword'
        - name: triggerType
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/assistant/agents:
    get:
      tags: [assistant]
      operationId: getAssistantAgents
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Success:
      description: Successful data envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
...
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing, expired, revoked, or invalid authentication
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 128 }
        password: { type: string, format: password, minLength: 1, maxLength: 256 }
...
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/LoginData' }
    ReauthRequest:
      type: object
      required: [password]
      properties:
        password: { type: string, format: password, minLength: 1, maxLength: 256 }
...